eMenu/requirements.txt
Django==3.2.12
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
drf-spectacular==0.21.2
celery==5.2.3
django-celery-results==2.2.0
django-celery-beat==2.2.1
coverage==6.3.2
model-bakery==1.4.0
Pillow==9.0.1
gunicorn==20.1.0